move functions to Functions folder
