+++ imports/api/routes/Routes.Private.js
@@ -3,7 +3,7 @@ import { Session } from 'meteor/session';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 import { BlazeLayout } from 'meteor/kadira:blaze-layout';
 
-import { wrs } from '/imports/framework/Functions.Async';
+import { wrs } from '/imports/framework/Functions/Async';
 import Helpers from '/imports/api/routes/Helpers';
 
 FlowRouter.route('/support', {
