+++ imports/api/routes/Routes.Public.js
@@ -4,7 +4,7 @@ import { Session } from 'meteor/session';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 import { BlazeLayout } from 'meteor/kadira:blaze-layout';
 
-import { wrs } from '/imports/framework/Functions.Async';
+import { wrs } from '/imports/framework/Functions/Async';
 import Helpers from '/imports/api/routes/Helpers';
 
 FlowRouter.notFound = {
