+++ imports/ui/calendar/calendar.js
@@ -5,7 +5,7 @@ import { TAPi18n } from 'meteor/tap:i18n';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 import moment from 'moment';
 
-import { wrs } from '/imports/framework/Functions.Async';
+import { wrs } from '/imports/framework/Functions/Async';
 
 import './calendar.tpl.jade';
 import './calendar.scss';
